Add CommonJS support and external TypeScript definitions.
This allows projects in environments such as Node, Webpack, Browserify,
and React Native to use WebCola. In addition, TypeScript projects will
be able to get WebCola type definitions in tsd 0.6 by running `tsd link`.

In general, to use WebCola, a project will modify package.json to
include the dependency. For example, to get WebCola from master, the
following would be added to `package.json`:

    "dependencies": {
      "cola": "tgdwyer/WebCola#master"
    }

Then, the user would run:

    npm install
    tsd link  # if TypeScript definitions were desired

Finally, the user would use WebCola as such:

    import cola = require("cola");
    // ...

The current TypeScript definitions are ambiant. The main implication
of that is that the import's name has to be 'cola'. This should likely
be improved in a future commit.