set node index before it's needed
fixes #134
edge source and target can be specified by integer index or by reference
to the node objects. if specified by reference, cola.js needs the node
objects to have an `index` member specifying the integer index.

it is already initializing this from the position in the array but it
needs to do this before calling the linkLengthCalculator; otherwise
lengths will not be applied.