fix(GridRouter): route sometimes crashs when node width are not equal
Problem: because center of grid line is calculated as avg position of
overlapping nodes, if those node doesn't have the same width, sometimes
the gridline is not intersected with node's bound. Thus, that node
doesn't have any ports.

-> `route` crash because GridRouter expects node has at least 1 port.

---

Solution: instead of using avg position of nodes, we could use the mid
point of overlap segment.